Implement iterated dominant set lookup
Apparently, using the initial weight calculated at the start of each
round to determine if a node belongs to the current dominant set is not
always sufficient, as the maximum induced node weight can be even
smaller than this limit.

Currently it seems that this occurs only in the last round, so checking
the segment size might work. Further investigation is needed.