Prepare handling of unassigned pixels
The issue mentioned in db4ec97 could occur only in the last round
indeed. More precisely, it occurs if and only if there are no
leftover pixels at the end, that is the size of the last segment
is at least (1 - TARGET_SEGMENTATION_RATIO) * pixelCount.

Actually, it's quite natural that in this case the induced node
weights are smaller than the initial weight of the round, as this
represents that the remaining pixels form a separate, indivisible
segment.